cleans up ui and does some other stuff too

+++ inference/db/tool_call_storage.py
@@ -208,7 +208,7 @@ class ToolCallStorage:
                         result_data=result_data,
                         error_message=row["error_message"],
                         execution_time_ms=(
-                            float(row["execution_time_ms"])
+                            int(row["execution_time_ms"])
                             if row["execution_time_ms"]
                             else None
                         ),
